rebased from main. Game logic & persistence

+++ src/main/java/ch/uzh/ifi/hase/soprafs24/websocket/dto/helper/WSCardContent.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public class WSCardContent {
-    private Long cardId;
+    private Integer cardId;
     private String songId;
     private String imageUrl;
 }
